Renaming scil_apply_transform_to_bvecs.py to scil_gradients_apply_transform.py
